Create Schwarzschild.sage
Schwarzschild.sage is a simple, elementary, first implementation of the Schwarzschild spacetime, Schwarzschild metric using Sage Manifolds 0.7 package for Sage Math. There are also answers/calculations for the Tutorial Sheet 13 Schwarzschild Spacetime for the WE Heraeus Gravity and Light International Winter School 2015.

At the end, there's an implementation of the Reissner-Nordstrom Geometry (RN) with the RN metric.