Create SO3.sage
Implementation of SO3 Lie Group and Lie Algebra as a manifold using Sage Math 6.6 and Sagemanifolds 0.7
Michał Bejger and Eric Gourgoulhon's Sagemanifolds 0.7 package is a delight to use and is excellent in that it is architected EXACTLY how differentiable manifolds behave.  
I follow Darryl D. Holm, Tanya Schmah, Cristina Stoica. Geometric mechanics and symmetry: From finite to infinite dimensions 2009 as a reference and liberally implement from there.
I'm going to put this same file and updates on ernestyalumni.wordpress.com and on my Patreon ernestyalumni (please go donate if you can to help support me!) because I want to be able to simply download the file as is and someone doesn't need to install github app or go through a command line prompt.